feat: implement game function, points, winner declarator, clean rounds
Implement a "game" function that uses five rounds (handled by
the "playRound" function) and also decides to declare who won the game.
Also, implement "points" as variables that are assigned to the user &
the NPC that help the game function reach this decision.
Also, clean up the rounds within the playRound function to accomodate
the case-insensitive logic made within an earlier commit in order to
remove redundant code.